리듀서 작성 : initialState를 구현한다
